requirements.txt
streamlit
pandas
plotly
firebase-admin
google-generativeai
